First fixes for Linux

+++ src/Net.cpp
@@ -1,3 +1,5 @@
+#include <cstdio>
+
 #include "Net.hpp"
 
 Net::Net(std::vector<uint16_t> layerSizes) : m_layerSizes(layerSizes) {}
@@ -5,8 +7,8 @@ Net::Net(std::vector<uint16_t> layerSizes) : m_layerSizes(layerSizes) {}
 void Net::printInfo()
 {
     printf("Number of nodes per layer:\n");
-    for (int layerIndex = 0; layerIndex < m_layerSizes.size(); ++layerIndex)
+    for (size_t layerIndex = 0; layerIndex < m_layerSizes.size(); ++layerIndex)
     {
-        printf("    %2d: %d\n", layerIndex, m_layerSizes[layerIndex]);
+        printf("    %2ld: %d\n", layerIndex, m_layerSizes[layerIndex]);
     }
 }
